+++ src/index.js
@@ -2,10 +2,12 @@ import processCSS from "./processCSS";
 import ContainerQuery from "./ContainerQuery";
 import withContainerQueryCSS from "./withContainerQueryCSS";
 import initialiseAllContainers from "./initialiseAllContainers";
+import initialiseAllContainersFromCSS from "./initialiseAllContainersFromCSS";
 
 export {
   processCSS,
   ContainerQuery,
   withContainerQueryCSS,
-  initialiseAllContainers
+  initialiseAllContainers,
+  initialiseAllContainersFromCSS
 };