+++ src/initialiseAllContainers.js
@@ -1,16 +1,16 @@
 import Container from "@zeecoder/container-query";
-
+import { SELECTOR } from "@zeecoder/container-query-meta-builder";
 /**
  * @param {{
  *   [selector]: Object,
  * }} meta
  */
 const initialiseAllContainers = meta => {
-  if (meta.selector) {
+  if (meta[SELECTOR]) {
     // Seems like a single container's stats object, so let's convert it to the
     // multi-container format.
     meta = {
-      [meta.selector]: meta
+      [meta[SELECTOR]]: meta
     };
   }